requirements.txt
h5py==2.7.1
Keras==2.0.6
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
protobuf==3.4.0
PyYAML==3.12
scipy==0.19.1
six==1.11.0
tensorflow-gpu==1.1.0
Theano==0.9.0
Werkzeug==0.12.2
